refactor(jobs): modularize jobs page and shift job details into modal
